refactor controllers and api routes for better modularity
